requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0

# Data storage
pyarrow>=14.0.0  # For parquet file support

# Time series modeling
statsmodels>=0.14.0  # For ARIMA
scikit-learn>=1.3.0  # For metrics and preprocessing
tensorflow>=2.15.0  # For LSTM models

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# Web framework
streamlit>=1.29.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
ipython>=8.0.0